Module Ox

module Action : sig ... end

The Action module defines a signature used to parameterise Learner.

module Config : sig ... end

The Config module defines the various configuration parameters for Learner.

module Dict : sig ... end

Definitions pertaining to the core data structure used to hold the population of classifiers.

module Environment : sig ... end

The Environment module is used to construct values of the environment to feed the Learner.

module Learner : sig ... end

The Learner module implements the XCS learning classifier system.

module Singlecore_dict : sig ... end

Dict implementation optimised for single core programs (or for OCaml < 5).

module Sensors : sig ... end

The Sensors module defines a signature used to parameterise Learner.