Module Ox.Learner

The Learner module implements the XCS learning classifier system.

module type S = sig ... end

Signature of a learner based on the XCS learning classifier system. Usage of this API always starts by creating a new learner via the S.create function. The API is imperative; the core functions (S.provide_environment, S.provide_intermediate_feedback, and S.provide_final_feedback modify the given learner). Until some termination condition is satisfied, the typical sequence of function calls depends on whether we are dealing with a single-step versus a multi-sep problem.

module Make_singlecore (Sensors_def : Sensors.DEF) (Action : Action.S) : S with type sensors = Sensors_def.sensors and type action = Action.t

module Make_multicore (Sensors_def : Sensors.DEF) (Action : Action.S) (MConfig : Multicore_config.S) : S with type sensors = Sensors_def.sensors and type action = Action.t