Module Ox

module Action : sig ... end

The Action module defines a signature used to parameterise Learner.

module Config : sig ... end

The Config module defines the various configuration parameters for Learner.

module Environment : sig ... end

The Environment module is used to construct values of the environment to feed the Learner.

module Learner : sig ... end

The Learner module implements the XCS learning classifier system.

module Multicore_config : sig ... end

Defines configuration options for multi-core Learners.

module Sensors : sig ... end

The Sensors module defines a signature used to parameterise Learner.